reconfigure file layout
