+++ .eslintrc.json
@@ -18,5 +18,7 @@
         "react"
     ],
     "rules": {
+        /* eslint-disable no-unused-vars */
+        /* eslint-disable react/prop-types */
     }
 }
