+++ src/pages/index.js
@@ -3,7 +3,7 @@ import React from 'react'
 
 import Image from 'next/image';
 
-import Layout from "../components/Layout"
+import Layout from "../components/Layout/Layout"
 
 import LandingStyle from "../styles/LandingStyle"
